[setup] linking to the original problem in the runs directory controlled by --linkproblem option

+++ python/piernik_setup.py
@@ -724,15 +724,15 @@ def setup_piernik(data=None):
         fatal_problem = True
 
     # Add a link to original problem directory to make it easier to access auxiliary files stored there
-    # ToDo: add a setup option to copy selected extra files
-    orig_prob_link = rundir + "problem"
-    if (os.path.islink(orig_prob_link)):
-        os.remove(orig_prob_link)
-    try:
-        os.symlink("../../" + probdir, orig_prob_link)
-    except FileExistsError:
-        print('\033[91m' + "Cannot create the link to original problem because '" + orig_prob_link + "' exists (and is not a symbolic link)" + '\033[0m')
-        fatal_problem = True
+    if options.link_problem:
+        orig_prob_link = rundir + "problem"
+        if (os.path.islink(orig_prob_link)):
+            os.remove(orig_prob_link)
+        try:
+            os.symlink("../../" + probdir, orig_prob_link)
+        except FileExistsError:
+            print('\033[91m' + "Cannot create the link to original problem because '" + orig_prob_link + "' exists (and is not a symbolic link)" + '\033[0m')
+            fatal_problem = True
 
     if (options.nocompile):
         print("\033[93mCompilation of '%s' skipped on request." % args[0] +
@@ -824,6 +824,9 @@ runs/<problem>_POSTFIX rather than runs/<problem> .""")
 
     parser.add_option("-k", "--keeppar", action="store_true", dest="keep_par",
                       help="Do not override existing problem.par file with the default one.")
+
+    parser.add_option("--linkproblem", action="store_true", dest="link_problem",
+                      help="Make a symbolic link to the original problem directory in the run directory.")
 
     if data is None:
         all_args = []
